refactor(callOrUse): spare a function creation
For our purpose, it's useless to return a function from `callOrUse`.
Additionally, with the current implementation, the "callOrUse" name is a
bit confusing because `callOrUse` doesn't call nor use the provided
function or object. It just wraps it in a function.

I was initially thinking of renaming this helper to something like
`wrap` but I figured we do not actually need to "wrap" anything.

So, let's stick with the "callOrUse" name and indeed call or use
the provided function or object.